chore: split respawn_guard tests into respawn_guard_tests.rs for the size guard
respawn_guard.rs crossed the 50KB size-guard byte limit (58798 bytes) after
the rework-marker tests landed. Move the test module to a sibling file
included via #[path], keeping it the same child module so private items and
use super::* semantics are unchanged. 413/1065 lines, 18KB/37KB — both under
the guard. All 33 respawn_guard tests still pass; clippy clean.
